ROUND3

+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/point/PointController.java
@@ -5,6 +5,8 @@ import com.loopers.application.point.PointInfo;
 import com.loopers.interfaces.api.ApiResponse;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
+
 @RestController
 public class PointController implements PointV1ApiSpec{
     private final PointFacade pointFacade;
@@ -26,7 +28,7 @@ public class PointController implements PointV1ApiSpec{
     @PostMapping("/api/v1/points/charge")
     @Override
     public ApiResponse<PointV1Dto.UserPointResponse> charge(@RequestHeader("X-USER-ID") String userId, @RequestBody PointV1Dto.ChargePointRequest request) {
-        long chargedPoint = pointFacade.charge(userId, request);
+        BigDecimal chargedPoint = pointFacade.charge(userId, request);
         return ApiResponse.success(new PointV1Dto.UserPointResponse(
                 userId,
                 chargedPoint
